Coupon test files restructured
