+++ phpunit.php
@@ -1,3 +1,4 @@
 <?php
 
 require 'vendor/autoload.php';
+require_once 'tests/includes/CartCouponTestCase.php';
